% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCoV.R
\name{MyCoV}
\alias{MyCoV}
\title{MyCoV takes a fasta file or multi-fasta file as input}
\usage{
MyCoV(fasta = NULL, temp_dir = NULL)
}
\arguments{
\item{fasta}{File name: the file containing the fasta-format sequences to be identified}

\item{temp_dir}{Folder name: the temporary folder in which to store files (eg. blast databse)}
}
\value{
A data frame summarising the BLAST results from the query
}
\description{
MyCoV takes a fasta file or multi-fasta file as input
}
