% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCoV.R
\name{plot_similarity}
\alias{plot_similarity}
\title{plot_similarity takes a mono-fasta file as an iput
It then uses 'ggtree' to provide a summary of how the queried sequence compared to classified sequences in our phylogenetic study.
It also provides a visual summary of host and country or origin metadata associated with the classified sequences.}
\usage{
plot_similarity(fasta = NULL, temp_dir = NULL)
}
\arguments{
\item{fasta}{file: a fasta file containing a single dna sequence in fasta format}

\item{temp_dir}{Folder name: the temporary folder in which to store files (eg. blast databse)}
}
\value{
A ggplot object summarising the data
}
\description{
Please remember to cite ggtree correctly if you go on to use this representation.
}
