% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCoV.R
\name{tabulate_CoV}
\alias{tabulate_CoV}
\title{tabulate_CoV takes the output of MyCoV as its input
It then uses 'formattable' to provide a summary of the classifications of the queried sequences.}
\usage{
tabulate_CoV(df = NULL)
}
\arguments{
\item{df}{data.frame: the output of the function MyCoV}
}
\value{
A 'formattable' object summarising the data
}
\description{
tabulate_CoV takes the output of MyCoV as its input
It then uses 'formattable' to provide a summary of the classifications of the queried sequences.
}
