% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCoV.R
\name{visualise_CoV}
\alias{visualise_CoV}
\title{visualise_CoV takes the output of MyCoV as its input
It then uses 'ggplot2' to provide a summary of the pairwise distances of the queried sequences to their closes BLAST hits in the context of all pairwise comparisons between Coronaviruses .}
\usage{
visualise_CoV(df = NULL)
}
\arguments{
\item{df}{data.frame: the output of the function MyCoV}
}
\value{
A ggplot object summarising the data
}
\description{
visualise_CoV takes the output of MyCoV as its input
It then uses 'ggplot2' to provide a summary of the pairwise distances of the queried sequences to their closes BLAST hits in the context of all pairwise comparisons between Coronaviruses .
}
